Use (custom) base-64 encoding for function name mangling
Once upon a time one developer claimed that a lot of bytes could be
saved by using base-64 encoding instead of base-36 encoding for mangling
function names with inline classes in parameter types. Now this is done,
and each name takes 1 byte less in the precious string table in the
class file, and there was much rejoicing.