Introduce notion of Outcome
Previously, 'Returns' was assumed to be an 'Effect' too, and conditional
clauses could have an 'Effect' on the LHS as a consequence.
This lead to various inconsistencies:
- 'contract { returns() }' was a correct contract (because each line was
expected to be 'Effect', so 'Returns' fit here)
- 'callsInPlace(...) implies ...' was a correct contract too (because
'CallsInPlace' is an 'Effect', so it fit to the LHS)

Now, we introduce notion of 'Outcome', which is:
- some outcome of a computation, which can be observed (like 'returns',
or, potentially, 'throws')
- doesn't form a statement
- 'Outcome' (and only 'Outcome') can appear on the LHS of a conditional
clause

'Returns' is an Outcome

Aside from solving inconsistencies mentioned above, it also opens ways
to clarify model and the implementation further.