Make EffectDeclarationInterpreter return Functor
Effectively, this commit just moves wrapping EsEffects into
SubstitutingFunctor from the call-site of 'tryInterpret' inside the body
of 'tryInterpret'.

This is closer to the proposed model (each "statement" of the contract
is interpreted into Functor), clarifies code a bit, but more
importantly, offers potential extensions more flexibility (i.e. if some
extension wouldn't work in terms of EsEffects and Substitutions, it
would be fine too)