Change compiler module dependencies scope in util-klib-metadata to 'compileOnly'.
When such dependencies have regular 'compile' scope, they're written to the pom file.
If someone depends on a kotlin-util-klib-metadata (in our case, this is kotlinx.team.infra plugin which serves internal purposes of a libraries team), transitive dependencies to 'frontend' and others become unresolved by Gradle, because this modules aren't published to repository separately.

Such dependencies have to be 'compileOnly', as it done in kotlin-native-library-reader.